move calculate_cumulative_distribution_in_place to bolts
